extracted common impl for ecies
