+++ components/crypto/crypto-ecies-core-impl/build.gradle
@@ -3,7 +3,7 @@ plugins {
     id 'corda.common-library'
 }
 
-description 'Crypto ECIES API'
+description 'Crypto ECIES core implemntation'
 
 dependencies {
     compileOnly 'org.osgi:osgi.annotation'
@@ -11,8 +11,11 @@ dependencies {
     implementation 'net.corda.kotlin:kotlin-stdlib-jdk8-osgi'
     implementation platform("net.corda:corda-api:$cordaApiVersion")
 
+    implementation project(":libs:crypto:crypto-core")
+
     api "net.corda:corda-cipher-suite"
 
+    api project(":components:crypto:crypto-ecies")
     api project(":libs:lifecycle:lifecycle")
 }
