+++ components/crypto/crypto-ecies-core-impl/src/main/java/net/corda/crypto/ecies/core/impl/package-info.java
@@ -1,0 +1,4 @@
+@Export
+package net.corda.crypto.ecies.core.impl;
+
+import org.osgi.annotation.bundle.Export;
