+++ components/crypto/crypto-ecies-impl/build.gradle
@@ -3,7 +3,7 @@ plugins {
     id 'corda.common-library'
 }
 
-description 'Crypto ECIES Implementation'
+description 'Crypto ECIES OSGi components'
 
 dependencies {
     compileOnly 'org.osgi:osgi.annotation'
@@ -16,7 +16,7 @@ dependencies {
 
     implementation project(":components:crypto:crypto-client")
     implementation project(":components:crypto:crypto-component-core-impl")
-    implementation project(":components:crypto:crypto-ecies")
+    implementation project(":components:crypto:crypto-ecies-core-impl")
     implementation project(":libs:crypto:crypto-core")
 
     testImplementation "org.assertj:assertj-core:$assertjVersion"
