+++ components/crypto/crypto-ecies-impl/src/main/kotlin/net/corda/crypto/ecies/impl/EphemeralKeyPairEncryptorImpl.kt
@@ -2,7 +2,7 @@ package net.corda.crypto.ecies.impl
 
 import net.corda.crypto.ecies.EncryptedDataWithKey
 import net.corda.crypto.ecies.EphemeralKeyPairEncryptor
-import net.corda.crypto.ecies.encryptWithEphemeralKeyPair
+import net.corda.crypto.ecies.core.impl.encryptWithEphemeralKeyPair
 import net.corda.v5.cipher.suite.CipherSchemeMetadata
 import org.osgi.service.component.annotations.Activate
 import org.osgi.service.component.annotations.Component
