+++ components/crypto/crypto-service-impl/build.gradle
@@ -23,6 +23,7 @@ dependencies {
     implementation project(":components:crypto:crypto-client")
     implementation project(":components:crypto:crypto-component-core-impl")
     implementation project(":components:crypto:crypto-ecies")
+    implementation project(":components:crypto:crypto-ecies-core-impl")
     implementation project(":components:crypto:crypto-persistence")
     implementation project(":components:crypto:crypto-service")
     implementation project(":libs:configuration:configuration-core")
@@ -33,6 +34,7 @@ dependencies {
 
     testImplementation "org.jetbrains.kotlin:kotlin-test:$kotlinVersion"
     testImplementation project(":components:crypto:crypto-component-test-utils")
+    testImplementation project(":components:crypto:crypto-ecies-core-impl")
     testImplementation project(":components:crypto:crypto-ecies-impl")
 
     testImplementation project(":components:crypto:crypto-db-persistence-model")
