+++ processors/crypto-processor/test.bndrun
@@ -17,6 +17,7 @@
     bnd.identity;id='net.corda.crypto-client-impl',\
     bnd.identity;id='net.corda.crypto-db-persistence-impl',\
     bnd.identity;id='net.corda.cipher-suite-impl',\
+    bnd.identity;id='net.corda.crypto-ecies-core-impl',\
     bnd.identity;id='net.corda.crypto-ecies-impl',\
     bnd.identity;id='net.corda.crypto-service-impl',\
     bnd.identity;id='net.corda.db-connection-manager-impl',\
