+++ settings.gradle
@@ -105,6 +105,7 @@ include 'components:crypto:crypto-component-test-utils'
 include 'components:crypto:crypto-db-persistence-impl'
 include 'components:crypto:crypto-db-persistence-model'
 include 'components:crypto:crypto-ecies'
+include 'components:crypto:crypto-ecies-core-impl'
 include 'components:crypto:crypto-ecies-impl'
 include 'components:crypto:crypto-persistence'
 include 'components:crypto:crypto-service'
