+++ tools/crypto-tck/build.gradle
@@ -13,9 +13,10 @@ dependencies {
     api 'net.corda.kotlin:kotlin-stdlib-jdk8-osgi'
     api platform("net.corda:corda-api:$cordaApiVersion")
 
+    implementation project(":components:crypto:crypto-ecies-core-impl")
+
     api "net.corda:corda-cipher-suite"
 
-    api project(":components:crypto:crypto-ecies")
     api project(':libs:crypto:cipher-suite-impl')
     api project(':libs:crypto:crypto-serialization-impl')
